requirements.txt
alembic==1.9.2
aniso8601==9.0.1
apispec==6.0.2
apispec-webframeworks==0.5.2
attrs==22.2.0
autopep8==2.0.1
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.0.1
click==8.1.3
cryptography==39.0.0
environs==9.5.0
Flask==2.2.2
flask-marshmallow==0.14.0
Flask-Migrate==4.0.2
flask-redoc==0.2.1
flask-restplus==0.13.0
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
idna==2.7
itsdangerous==2.1.2
Jinja2==3.1.2
jsonmerge==1.9.0
jsonschema==4.17.3
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.28.1
packaging==23.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pycparser==2.21
pyOpenSSL==23.0.0
pyrsistent==0.19.3
python-dotenv==0.21.1
pytz==2022.7.1
PyYAML==6.0
requests==2.20.0
six==1.16.0
SQLAlchemy==1.4.46
tomli==2.0.1
urllib3==1.24.3
Werkzeug==2.2.2
